refactor(control-ui): split static response helpers
